+++ Lib/fontbakery/profiles/opentype.py
@@ -71,7 +71,6 @@ OPENTYPE_PROFILE_CHECKS = [
     'com.google.fonts/check/maxadvancewidth',
     'com.google.fonts/check/points_out_of_bounds',
     'com.google.fonts/check/glyf_non_transformed_duplicate_components',
-    'com.google.fonts/check/all_glyphs_have_codepoints',
     'com.google.fonts/check/code_pages',
     'com.google.fonts/check/layout_valid_feature_tags',
     'com.google.fonts/check/layout_valid_script_tags',